% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cells_to_group10X.R
\name{cells_to_group10X}
\alias{cells_to_group10X}
\title{Grouping spots pairs into groups for 10X datasets}
\usage{
cells_to_group10X(spatial, ident, group_size, celltype = NULL)
}
\arguments{
\item{spatial}{spatial coordinate information, a dataframe which rows represent spots and cols represent coordinate}

\item{ident}{celltype information, a dataframe which rows represent spots and cols represent celltype}

\item{group_size}{The group size parameters determined by humans. For single-cell spatial RNA-seq data, we recommend grouping them into 5000 spots, while for spatial bulk RNA-seq data, grouping into 2000 spots is recommended.}
}
\value{
grouping results in spot pairs, each group of 5 rows represents:
spatial distance between spot pairs, index of the spot sending the signal, index of the spot receiving the signal, cell type of the spot sending the signal, cell type of the spot receiving the signal.
}
\description{
Grouping spots pairs into groups for 10X datasets
}
\examples{
cells_group = cells_to_group10X(spatial, ident, group_size)
}
