% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cccM.R
\name{get_cccM}
\alias{get_cccM}
\title{Get the matrix contain every spots' CCC feature}
\usage{
get_cccM(result, cells_group, spatial, gene, match_list1, match_list2)
}
\arguments{
\item{result}{result get from function IGAN}

\item{cells_group}{grouping result from function cells_to_group or cells_to_group10X}

\item{spatial}{spatial coordinate information, a dataframe which rows represent spots and cols represent coordinate}

\item{gene}{gene name of the dataset}

\item{match_list1}{gene list selected in sending spots}

\item{match_list2}{gene list selected in receiving spots}
}
\value{
a matrix contain every spots' CCC feature
}
\description{
Get the matrix contain every spots' CCC feature
}
\examples{
cccM = get_cccM(result, cells_group, spatial, gene, match_list1, match_list2)
}
