% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IGAN.R
\name{IGAN}
\alias{IGAN}
\title{Compute gene-gene associations in every single spots pair}
\usage{
IGAN(match_list1, match_list2, gene, cells_group, M, p_value)
}
\arguments{
\item{match_list1}{gene list selected in sending spots}

\item{match_list2}{gene list selected in receiving spots}

\item{gene}{gene name of the dataset}

\item{cells_group}{grouping result from function cells_to_group or cells_to_group10X}

\item{M}{gene expression matrix, a dataframe which cols represent genes and rows represent spots}

\item{p_value}{p_value used in Statistical Testing}
}
\description{
Compute gene-gene associations in every single spots pair
}
\examples{
result = IGAN(match_list1, match_list2, gene, cells_group, M, p_value)
}
