% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cellchat.R
\name{plot_cellchat}
\alias{plot_cellchat}
\title{Get the cellchat's network circle graph}
\usage{
plot_cellchat(useGene, result, cells_group, gene, match_list1, ident)
}
\arguments{
\item{useGene}{the gene's name used to plot}

\item{result}{result get from function IGAN}

\item{cells_group}{grouping result from function cells_to_group or cells_to_group10X}

\item{gene}{gene name of the dataset}

\item{match_list1}{gene list selected in sending spots}

\item{ident}{celltype information, a dataframe which rows represent spots and cols represent celltype}
}
\value{
a cellchat's network circle graph
}
\description{
Get the cellchat's network circle graph
}
\examples{
plot_cellchat('SPP1',result,cells_group,gene,match_list1,ident)
}
