% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cor.R
\name{plot_cor}
\alias{plot_cor}
\title{Get the graph of CCC activity of every spot}
\usage{
plot_cor(result, cells_group, spatial)
}
\arguments{
\item{result}{result get from function IGAN}

\item{cells_group}{grouping result from function cells_to_group or cells_to_group10X}

\item{spatial}{spatial coordinate information, a dataframe which rows represent spots and cols represent coordinate}
}
\value{
graph of CCC activity of every spot
}
\description{
Get the graph of CCC activity of every spot
}
\examples{
a = plot_cor(result, cells_group, spatial)
          a + scale_color_distiller(palette = "Set1")
}
