% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Genecor.R
\name{plot_Genecor}
\alias{plot_Genecor}
\title{Get the graph of useGene's CCC activity of every spot}
\usage{
plot_Genecor(useGene, result, cells_group, spatial, match_list1, gene)
}
\arguments{
\item{useGene}{the gene's name used to plot}

\item{result}{result get from function IGAN}

\item{cells_group}{grouping result from function cells_to_group or cells_to_group10X}

\item{spatial}{spatial coordinate information, a dataframe which rows represent spots and cols represent coordinate}

\item{match_list1}{gene list selected in sending spots}

\item{gene}{gene name of the dataset}
}
\value{
graph of useGene's CCC activity of every spot
}
\description{
Get the graph of useGene's CCC activity of every spot
}
\examples{
a = plot_Genecor('SPP1', result, cells_group, spatial, match_list1, gene)
          a + scale_color_distiller(palette = "Set1")
}
