% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ligand_target.R
\name{plot_ligand_target}
\alias{plot_ligand_target}
\title{Title}
\usage{
plot_ligand_target(
  result,
  spatial,
  cells_group,
  gene,
  match_list1,
  match_list2,
  genename
)
}
\arguments{
\item{genename}{}
}
\description{
Title
}
