% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pattern.R
\name{plot_pattern}
\alias{plot_pattern}
\title{Get the graph of CCC pattern}
\usage{
plot_pattern(
  result,
  cells_group,
  gene,
  match_list1,
  match_list2,
  ident,
  pattern,
  n
)
}
\arguments{
\item{result}{result get from function IGAN}

\item{cells_group}{grouping result from function cells_to_group or cells_to_group10X}

\item{gene}{gene name of the dataset}

\item{match_list1}{gene list selected in sending spots}

\item{match_list2}{gene list selected in receiving spots}

\item{ident}{celltype information, a dataframe which rows represent spots and cols represent celltype}

\item{pattern}{set it's sending pattern or receiving pattern}

\item{n}{the number of pattern}
}
\value{
graph of CCC pattern
}
\description{
Get the graph of CCC pattern
}
\examples{
pattern_graph = plot_pattern(result,cells_group,gene,match_list1,match_list2,ident,'outgoing',4)
}
