% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sankey.R
\name{plot_sankey}
\alias{plot_sankey}
\title{Get the  sankey graph of ligand's GO-ligand-recepter-downstream pathway}
\usage{
plot_sankey(
  result,
  cells_group,
  gene,
  match_list1,
  match_list2,
  p,
  minGenes,
  OrgDb,
  send_cell = NULL,
  rec_cell = NULL,
  threshold
)
}
\arguments{
\item{result}{result get from function IGAN}

\item{cells_group}{grouping result from function cells_to_group or cells_to_group10X}

\item{gene}{gene name of the dataset}

\item{match_list1}{gene list selected in sending spots}

\item{match_list2}{gene list selected in receiving spots}

\item{p}{p_value used in KEGG enriched analysis}

\item{minGenes}{the min number of associated genes in ligands' downstream pathways}
}
\value{
a sankey graph of ligand's GO-ligand-recepter-downstream pathway
}
\description{
Get the  sankey graph of ligand's GO-ligand-recepter-downstream pathway
}
\examples{
sankey_graph = plot_sankey(result,cells_group,gene,match_list1,match_list2,0.001,5)
}
